04_Production_RAG/requirements.txt
python-dotenv
nest-asyncio
langchain
langchain-community
langchain-openai
qdrant-client
pandas
langsmith
openai
